[#25] Massive rebase commit. Fixes all data sets. May have compatibility issues. Currently under heavy testing

+++ src/coins.cpp
@@ -9,6 +9,7 @@
 #include "random.h"
 
 #include <assert.h>
+#include <boost/foreach.hpp>
 
 bool CCoinsView::GetCoin(const COutPoint &outpoint, Coin &coin) const { return false; }
 uint256 CCoinsView::GetBestBlock() const { return uint256(); }
@@ -249,14 +250,14 @@ double CCoinsViewCache::GetPriority(const CTransaction &tx, int nHeight, CAmount
         return 0.0;
     double dResult = 0.0;
     BOOST_FOREACH(const CTxIn& txin, tx.vin)
-    {
-        const Coin& coin = AccessCoin(txin.prevout);
-        if (coin.IsSpent()) continue;
-        if (coin.nHeight <= nHeight) {
-            dResult += coin.out.nValue * (nHeight-coin.nHeight);
-            inChainInputValue += coin.out.nValue;
-        }
-    }
+                {
+                    const Coin& coin = AccessCoin(txin.prevout);
+                    if (coin.IsSpent()) continue;
+                    if (coin.nHeight <= nHeight) {
+                        dResult += (double)coin.out.nValue * (nHeight-coin.nHeight);
+                        inChainInputValue += coin.out.nValue;
+                    }
+                }
     return tx.ComputePriority(dResult);
 }
